requirements.txt
aiofiles==24.1.0
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema_pydantic==0.6
jsonschema-specifications==2025.4.1
lark==1.2.2
marshmallow==3.26.1
motor==3.7.1
networkx==3.5
openapi-pydantic==0.5.1
orjson==3.10.18
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
PyYAML==6.0.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
scarf-sdk==0.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.1
sympy==1.14.0
tenacity==8.5.0
TikTokApi==7.1.0
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
